% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{cartesian_distance}
\alias{cartesian_distance}
\title{Cartesian distance between two points}
\usage{
cartesian_distance(x1, y1, x2, y2)
}
\arguments{
\item{x1}{double, X coordinate of the first point}

\item{y1}{double, Y coordinate of the first point}

\item{x2}{double, X coordinate of the second point}

\item{y2}{double, Y coordinate of the second point}
}
\value{
a numeric value of class double
}
\description{
Given two planar coordinates, estimates the distance between them
}
