% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fil_wt_sf.R
\name{fil_wt_sf}
\alias{fil_wt_sf}
\title{Spatial smoothing with Whittaker-Eilers filter.}
\usage{
fil_wt_sf(input, lamda = 10, order = 2, keep_start = TRUE)
}
\arguments{
\item{input}{\code{sf} object of geometry type LINESTRING. Note that
currently, only single lines are supported.}

\item{lamda}{Double, inherits from \code{\link[=fil_wt_x]{fil_wt_x()}}.}

\item{order}{Double, inherits from \code{\link[=fil_wt_x]{fil_wt_x()}}.}

\item{keep_start}{Boolean, should starting/ending points remained unchanged?
Default \code{TRUE}}
}
\value{
an \code{sf} object
}
\description{
Applies \code{\link[=fil_wt_x]{fil_wt_x()}} for \code{sf} objects.
Since the majority of linear objects (i.e., rivers, roads, etc.)
have non-equally spaced coordinates, the vector of sample positions
(see \href{https://www.anbowell.com/blog/the-perfect-way-to-smooth-your-noisy-data}{this blog post}
for details) is estimated automatically based on the distance between
point locations. Depending on planar or angular geometry, the Cartesian or
Haversine distance is calculated.
}
\examples{
if (rlang::is_installed("spData")) {
library(sf)
library(filtrs)
library(spData)

rivers <-
  seine[3, ]

rivers_smoothed <-
  fil_wt_sf(rivers, lamda = 10e3, order = 3)

plot(
  sf::st_geometry(rivers),
  col = "grey30",
  lwd = 3.5,
  axes = TRUE
)
plot(
  sf::st_geometry(rivers_smoothed),
  col = 'firebrick3',
  lwd = 2,
  add = TRUE
)

# Add the legend
legend(
  "bottomright",
  legend = c("Original", "Smoothed"),
  col = c("grey30", "firebrick3"),
  lwd = c(3, 3)
)
}
}
