% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fil_wt_x}
\alias{fil_wt_x}
\title{Perform a Whittaker-Eilers smoother for non-equally spaced data.}
\usage{
fil_wt_x(y, x, lambda, order)
}
\arguments{
\item{y}{numeric vector to smooth (must be of class double!!)}

\item{x}{numeric vector with sample positions}

\item{lambda}{an number of class double for smoothing control. The larger the lambda, the smoother the data.}

\item{order}{an integer for smoothing control. A higher order means the Whittaker will consider more adjacent elements while smoothing.}
}
\description{
Insanely fast and reliable smoothing and interpolation with the Whittaker-Eilers method for non-equally spaced data.
}
\examples{
data("airquality")

airquality$Ozone_smooth <- 
 fil_wt_x(as.double(airquality$Ozone), as.double(seq_along(airquality$Ozone)), 10, 2)

plot(airquality$Ozone, type = "l", xlab = "Days", ylab = "Ozone")
lines(airquality$Ozone_smooth, co = "red")
}
