% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{filtrs-package}
\alias{filtrs}
\alias{filtrs-package}
\title{filtrs: Rust-boosted linear and spatial filtering in R}
\description{
This package offers quick smoothing and linear filtering of spatial and timeseries objects.
}
\author{
\strong{Maintainer}: Anatoly Tsyplenkov \email{atsyplenkov@gmail.com} (\href{https://orcid.org/0000-0003-4144-8402}{ORCID})

}
