% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{haversine_distance}
\alias{haversine_distance}
\title{Harvesine distance between two points}
\usage{
haversine_distance(lon1, lat1, lon2, lat2)
}
\arguments{
\item{lon1}{double, longitude (X1) of the first point}

\item{lat1}{double, latitude (Y1) of the first point}

\item{lon2}{double, longitude (X2) of the second point}

\item{lat2}{double, latitude (Y2) of the second point}
}
\value{
a numeric value of class double
}
\description{
Given two angular coordinates, estimates the distance between them
}
