% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{filtrs-package}
\alias{filtrs}
\alias{filtrs-package}
\title{filtrs: Rust-boosted linear and spatial filtering in R}
\description{
This package offers quick smoothing and linear filtering of spatial and timeseries objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/atsyplenkov/filtrs}
  \item Report bugs at \url{https://github.com/atsyplenkov/filtrs/issues}
}

}
\author{
\strong{Maintainer}: Anatoly Tsyplenkov \email{atsyplenkov@fastmail.com} (\href{https://orcid.org/0000-0003-4144-8402}{ORCID})

}
